DB/Spawns: Fix previous commit (512aa8897bf30bbb8)

+++ src/server/game/Entities/Player/Player.cpp
@@ -4822,7 +4822,7 @@ void Player::DeleteFromDB(uint64 playerguid, uint32 accountId, bool updateRealmC
                     Field* fields = resultMail->Fetch();
 
                     uint32 mail_id       = fields[0].GetUInt32();
-                    uint16 mailType      = fields[1].GetUInt8();
+                    uint8 mailType       = fields[1].GetUInt8();
                     uint16 mailTemplateId= fields[2].GetUInt16();
                     uint32 sender        = fields[3].GetUInt32();
                     std::string subject  = fields[4].GetString();
